Code review: - harden relay backoff fields for cross-thread access - reuse EmptyConnectionListener in backoff test - extract shared relay-client test fakes
